+++ tests/http/server.cpp
@@ -115,6 +115,8 @@ BOOST_AUTO_TEST_CASE(construction)
     BOOST_CHECK_NE( server1.getURI().getHost(), "" );
     BOOST_CHECK_NE( server1.getURI().getHost(), "*" );
     BOOST_CHECK_NE( server1.getURI().getPort(), 0 );
+    BOOST_CHECK_NO_THROW( server1.getSocketDescriptor( ));
+    BOOST_CHECK_GT( server1.getSocketDescriptor(), 0 );
 
     const zeq::URI uri( "tcp://" );
     zeq::http::Server server2( uri );
@@ -124,6 +126,8 @@ BOOST_AUTO_TEST_CASE(construction)
 
     BOOST_CHECK_THROW( zeq::http::Server( server2.getURI( )),
                        std::runtime_error );
+    BOOST_CHECK_NO_THROW( server2.getSocketDescriptor( ));
+    BOOST_CHECK_GT( server1.getSocketDescriptor(), 0 );
 }
 
 BOOST_AUTO_TEST_CASE(construction_argv_host_port)